220: moving some classes, fixing tests
