+++ src/main/java/tech/units/indriya/function/NumberSystem.java
@@ -27,7 +27,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package tech.units.indriya.internal.function.calc;
+package tech.units.indriya.function;
 
 /**
  * Provides basic {@link Number} operations.
